dm vdo: move vdo size constants to encodings.c
Move static constants from encodings.h to encodings.c to
fix upstream compile warnings about unused constants.
Upstream these constants are only use in encoding.c. However,
these constants are also used in userspace tools and some
tests, so internally it's easiest to leave the constants
in encodings.h. Add an assertion in encodings.c so that
we'll notice the mismatch if we ever change the value of
these constants.

This is an edit to Susan's earlier commit 9462281cde95 and
will be squashed into that one when going upstream.

Signed-off-by: Matthew Sakai <msakai@redhat.com>