dm vdo: internalize UDS index size computation
David Alan Gilbert, upstream, wants to remove this function because
it is not currently used in the kernel. However, UDS unit tests use
it, as well as our userspace tools, so instead we will hide it from
the kernel.

Cf. 29e11586b56a ("dm vdo: Remove unused uds_compute_index_size")

Note that this function will be needed by the vdoformat-in-kernel
story, but in order to make upstreaming that story easier, we
should incorporate this removal now.

Signed-off-by: Matthew Sakai <msakai@redhat.com>